requirements.txt
fasttext
pytorch-crf
gensim
pytorch